+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "./components/Header";
 import ErrorBoundary from "./components/ErrorBoundary";
-import MainPage from "./components/Pages/MainPage";
+import MainPage from "./Pages/MainPage";
 import Footer from "./components/Footer";
 import StyledApp from "./StyledApp";
