+++ src/components/MoviesList/MoviesList.type.ts
