+++ src/components/Sorter/Sorter.type.ts
