+++ src/components/StatusBar/StatusBar.type.ts
