+++ src/components/StatusBar/index.ts
